-Fixed the Sector Manager (Missing the hashcode function for SectorLocations) -Added local entities to client -Added entity packets
TODO:
-Fix up the primary sending of entities
-Add graphical info to entity defs
-Finish up seamless rendering loop

+++ Common/src/com/pi/common/database/GraphicsAnimation.java
@@ -26,7 +26,7 @@ public class GraphicsAnimation extends GraphicsObject {
     }
 
     @Override
-    public void setPosition(int x, int y, int width, int height) {
+    public void setPosition(float x, float y, float width, float height) {
 	super.setPosition(x, y, width, height);
 	this.frameWidth = (int) (super.getPositionWidth() / frames);
     }
